Added first real model test. (attempt 66)

+++ RFNN/trainnonlin/mod_training.py
@@ -47,7 +47,7 @@ def apply_model():
     new_saver = tf.train.import_meta_graph('/results/weights.meta')
     new_saver.restore(sess, tf.train.latest_checkpoint('/results'))
     
-    return model(np.random.random(5), weights)
+    return model(np.random.random(5, dtype = 'float32'), weights)
             
             
 def model(data, weights):
